% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/advocate_case.R
\docType{data}
\name{advocate_case}
\alias{advocate_case}
\title{Advocate bio data for HCA cases 1995-2019}
\format{
\subsection{\code{advocate_case}}{

A tibble with 11,423 rows and 15 columns:
\describe{
\item{justice}{HCDB justice ID number}
\item{caseNumber}{HCA docket number}
\item{caseName}{name of the case}
...
}
}
}
\source{
Robinson, Z., Leslie, P. (High Court Database)
}
\usage{
advocate_case
}
\description{
Biographical data from all representatives appearing before the High Court of Australia in full hearings, 1995-2019
}
\keyword{datasets}
