% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attorney_case.R
\docType{data}
\name{attorney_case}
\alias{attorney_case}
\title{Attorney bio data for HCA cases 1995-2019}
\format{
\subsection{\code{justice_bio}}{

A tibble with 1755 rows and 436 columns:
\describe{
\item{justice}{HCDB justice ID number}
\item{caseNumber}{HCA docket number}
\item{caseName}{name of the case}
...
}
}
}
\source{
Robinson, Z., Leslie, P. (High Court Database)
}
\usage{
attorney_case
}
\description{
Biographical data from all attorneys appearing before the High Court of Australia in full hearings, 1995-2019
}
\keyword{datasets}
