% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcdb-package.R
\docType{package}
\name{hcdb-package}
\alias{hcdb}
\alias{hcdb-package}
\title{hcdb: An R package for the High Court Database}
\description{
Provides easy access to the Australian High Court Database (HCDB), a structured dataset encoding all full judgments of the High Court of Australia from 1995 to 2020 and all special leave applications from 2003 to 2018. For details see Leslie et al. (2026) \doi{10.1177/2755323X261438254}.
}
\details{
The hcdb package provides four datasets. Each has a corresponding codebook
vignette with full variable documentation:\tabular{ll}{
   Dataset \tab Codebook \cr
   \link{justice_decision} \tab \code{vignette("justice_decision_codebook")} \cr
   \link{justice_decision} \tab \code{vignette("justice_decision_core_codebook")} \cr
   \link{special_leave} \tab \code{vignette("special_leave_codebook")} \cr
   \link{justice_bio} \tab \code{vignette("justice_bio_codebook")} \cr
   \link{advocate_case} \tab \code{vignette("advocate_case_codebook")} \cr
}


To browse all codebooks at once: \code{browseVignettes("hcdb")}
}
\references{
Leslie, P., Robinson, Z., Smyth, R., & Jacobi, T. (2026). Supporting comparative
studies of judicial behavior: Introducing the Australian High Court Database. \emph{Journal of
Law & Empirical Analysis}, \emph{3}(1), 251–263. \doi{10.1177/2755323X261438254}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/palesl/hcdb}
  \item Report bugs at \url{https://github.com/palesl/hcdb/issues}
}

}
\author{
\strong{Maintainer}: Pat Leslie \email{patrick.leslie@anu.edu.au} (\href{https://orcid.org/0000-0002-9218-8981}{ORCID})

Authors:
\itemize{
  \item Pat Leslie \email{patrick.leslie@anu.edu.au} (\href{https://orcid.org/0000-0002-9218-8981}{ORCID})
}

}
\keyword{internal}
