% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justice_bio.R
\docType{data}
\name{justice_bio}
\alias{justice_bio}
\title{justice bio data, 1995-2021}
\format{
\subsection{\code{justice_bio}}{

A tibble with 23 rows and 155 columns:
\describe{
\item{justice}{HCDB justice ID number}
\item{justiceName}{Justice names}
\item{dateNom}{Date of nomination}
...
}
}
}
\source{
Robinson, Z., Leslie, P. (High Court Database)
}
\usage{
justice_bio
}
\description{
Biographical data from all justices appointed to the High Court of Australia, 1995-2021
}
\references{
Leslie, P., Robinson, Z., Smyth, R., & Jacobi, T. (2026). Supporting comparative
studies of judicial behavior: Introducing the Australian High Court Database. \emph{Journal of
Law & Empirical Analysis}, \emph{3}(1), 251–263. \doi{10.1177/2755323X261438254}
}
\seealso{
Full variable documentation: \code{vignette("justice_bio_codebook", package = "hcdb")}
}
\keyword{datasets}
