% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justice_decision.R
\docType{data}
\name{justice_decision}
\alias{justice_decision}
\title{justice decision data, 1995-2020}
\format{
\subsection{\code{justice_decision}}{

A tibble with 12,726 rows and 319 columns:
\describe{
\item{HCDBcaseId}{HCDB case ID number}
\item{clrCite}{Case Law Research citation}
\item{aljrCite}{Australian Law Journal Reports citation}
...
}
}
}
\source{
Robinson, Z., Leslie, P. (High Court Database)
}
\usage{
justice_decision
}
\description{
Data from all justice decisions in all cases in the High Court of Australia, 1995-2020
}
\references{
Leslie, P., Robinson, Z., Smyth, R., & Jacobi, T. (2026). Supporting comparative
studies of judicial behavior: Introducing the Australian High Court Database. \emph{Journal of
Law & Empirical Analysis}, \emph{3}(1), 251–263. \doi{10.1177/2755323X261438254}
}
\seealso{
Full variable documentation: \code{vignette("justice_decision_codebook", package = "hcdb")},
\code{vignette("justice_decision_core_codebook", package = "hcdb")}
}
\keyword{datasets}
