% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special_leave.R
\docType{data}
\name{special_leave}
\alias{special_leave}
\title{Special Leave Data, 2003-2018}
\format{
\subsection{\code{special_leave}}{

A tibble: 8,412 × 48:
\describe{
\item{HCSLDBId}{High Court Special Leave DB case ID}
\item{caseName}{Case Names}
\item{caseNumber}{High Court Docket Number}
...
}
}
}
\source{
Robinson, Z., Leslie, P. (High Court Database)
}
\usage{
special_leave
}
\description{
Special Leave Data, 2003-2018
}
\keyword{datasets}
