% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justice_bio.R
\docType{data}
\name{justice_bio}
\alias{justice_bio}
\title{justice bio data, 1995-2021}
\format{
\subsection{\code{justice_bio}}{

A tibble with 23 rows and 155 columns:
\describe{
\item{justiceID}{HCDB justice ID number}
\item{justiceName}{Justice names}
\item{dateNom}{Date of nomination}
...
}
}
}
\source{
Robinson, Z., Leslie, P. (High Court Database)
}
\usage{
justice_bio
}
\description{
Biographical data from all justices appointed to the High Court of Australia, 1995-2021
}
\keyword{datasets}
