% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justice_decision.R
\docType{data}
\name{justice_decision}
\alias{justice_decision}
\title{justice decision data, 1995-2020}
\format{
\subsection{\code{justice_decision}}{

A tibble with 12,726 rows and 328 columns:
\describe{
\item{HCDBcaseId}{HCDB case ID number}
\item{clrCite}{Case Law Research citation}
\item{aljrCite}{Australian Law Journal Reports citation}
...
}
}
}
\source{
Robinson, Z., Leslie, P. (High Court Database)
}
\usage{
justice_decision
}
\description{
Data from all justice decisions in all cases in the High Court of Australia, 1995-2020
}
\keyword{datasets}
